--- src/input4mips_validation/main.py
@@ -1,7 +1,0 @@
-"""
-Command-line interface entry point
-"""
-from input4mips_validation.cli.root import app
-
-if __name__ == "__main__":  # pragma: no cover
-    app()
